web/dashboard/backend/requirements.txt
websockets
python-dotenv
stellar-sdk
